linux: Change the type for `Window` to `c_ulong`

+++ src/linux.rs
@@ -2,12 +2,12 @@ use libc;
 
 use crate::{Key, KeyboardControllable, MouseButton, MouseControllable};
 
-use libc::{c_char, c_int, c_void, useconds_t};
+use libc::{c_char, c_int, c_ulong, c_void, useconds_t};
 use std::{borrow::Cow, ffi::CString, ptr};
 
-const CURRENT_WINDOW: c_int = 0;
+const CURRENT_WINDOW: c_ulong = 0;
 const DEFAULT_DELAY: u64 = 12000;
-type Window = c_int;
+type Window = c_ulong;
 type Xdo = *const c_void;
 
 #[link(name = "xdo")]
